Fix flaky mails test when real http reqeust to outsourced API was used
