+++ src/main/java/pl/newsler/api/IELAMailController.java
@@ -10,18 +10,18 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import pl.newsler.commons.exception.InvalidUserDataException;
 import pl.newsler.components.emaillabs.exception.InvalidDateException;
 import pl.newsler.components.emaillabs.usecase.ELAGetMailResponse;
-import pl.newsler.components.emaillabs.usecase.ELAMailSendRequest;
-import pl.newsler.components.emaillabs.usecase.ELAMailScheduleRequest;
+import pl.newsler.components.emaillabs.usecase.ELAInstantMailRequest;
+import pl.newsler.components.emaillabs.usecase.ELAScheduleMailRequest;
 
 import java.util.List;
 
 @RequestMapping(NLApi.V1 + "/api/mails")
 public interface IELAMailController {
     @PostMapping("/instant")
-    ResponseEntity<HttpStatus> queueAndExecute(@RequestBody ELAMailSendRequest request) throws InvalidUserDataException;
+    ResponseEntity<HttpStatus> queueAndExecute(@RequestBody ELAInstantMailRequest request) throws InvalidUserDataException;
 
-    @PostMapping("/scheduled")
-    ResponseEntity<HttpStatus> schedule(@RequestBody ELAMailScheduleRequest request) throws InvalidUserDataException, InvalidDateException;
+    @PostMapping("/schedule")
+    ResponseEntity<HttpStatus> schedule(@RequestBody ELAScheduleMailRequest request) throws InvalidUserDataException, InvalidDateException;
 
     @GetMapping("/{userId}")
     ResponseEntity<List<ELAGetMailResponse>> fetchAllMails(@PathVariable("userId") String userId) throws InvalidUserDataException;
