+++ src/main/java/pl/newsler/components/emaillabs/IELAMailService.java
@@ -4,15 +4,15 @@ import pl.newsler.commons.exception.InvalidUserDataException;
 import pl.newsler.commons.model.NLUuid;
 import pl.newsler.components.emaillabs.exception.InvalidDateException;
 import pl.newsler.components.emaillabs.usecase.ELAGetMailResponse;
-import pl.newsler.components.emaillabs.usecase.ELAMailSendRequest;
-import pl.newsler.components.emaillabs.usecase.ELAMailScheduleRequest;
+import pl.newsler.components.emaillabs.usecase.ELAInstantMailRequest;
+import pl.newsler.components.emaillabs.usecase.ELAScheduleMailRequest;
 
 import java.util.List;
 
 public interface IELAMailService {
-    void queue(ELAMailSendRequest request) throws InvalidUserDataException;
+    void queue(ELAInstantMailRequest request) throws InvalidUserDataException;
 
-    void schedule(ELAMailScheduleRequest request) throws InvalidUserDataException, InvalidDateException;
+    void schedule(ELAScheduleMailRequest request) throws InvalidUserDataException, InvalidDateException;
 
     List<ELAGetMailResponse> fetchAllMails(NLUuid userId) throws InvalidUserDataException;
 }
