+++ src/test/java/pl/newsler/testcommons/InMemoryJpaRepository.java
@@ -1,6 +1,7 @@
 package pl.newsler.testcommons;
 
 import lombok.RequiredArgsConstructor;
+import org.jetbrains.annotations.NotNull;
 import org.springframework.data.domain.Example;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -9,6 +10,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.query.FluentQuery;
 import org.springframework.security.crypto.password.PasswordEncoder;
 
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -83,8 +85,13 @@ public class InMemoryJpaRepository<T, ID> implements JpaRepository<T, ID> {
     }
 
     @Override
-    public <S extends T> List<S> saveAll(Iterable<S> entities) {
-        throw new PleaseImplementMeException();
+    public <S extends T> @NotNull List<S> saveAll(Iterable<S> entities) {
+        final List<S> list = new ArrayList<>();
+        for (final S entity : entities) {
+            list.add(save(entity));
+        }
+
+        return list;
     }
 
     @Override
